Change Contours size to be `f64` instead of `usize`

+++ plotly/src/traces/contour.rs
@@ -53,7 +53,7 @@ pub struct Contours {
     r#type: Option<ContoursType>,
     start: Option<f64>,
     end: Option<f64>,
-    size: Option<usize>,
+    size: Option<f64>,
     coloring: Option<Coloring>,
     #[serde(rename = "showlines")]
     show_lines: Option<bool>,
@@ -522,7 +522,7 @@ mod tests {
             .type_(ContoursType::Levels)
             .start(0.0)
             .end(10.0)
-            .size(5)
+            .size(5.0)
             .coloring(Coloring::HeatMap)
             .show_lines(true)
             .show_labels(false)
@@ -535,7 +535,7 @@ mod tests {
             "type": "levels",
             "start": 0.0,
             "end": 10.0,
-            "size": 5,
+            "size": 5.0,
             "coloring": "heatmap",
             "showlines": true,
             "showlabels": false,
